Reconcile array-transfer-phase3 (#1470) and reverse (#1274) onto embedding
Full file-by-file conflict resolution of the squash-merge, plus the
verification/fix pass that followed: 1470's object model/perf work
adapted to reverse's per-context (no ambient JPContext_global) ownership
rules throughout the native tree, JPRef replaced by the two established
patterns (context-owned bare globals released in detachJVM(), or
GlobalPool-backed jref handles for dynamically created wrapper objects),
and JPContext_global itself removed entirely now that every call site has
a real context threaded in from its caller instead of an ambient global
that silently returns the wrong sub-interpreter's context.

Along the way: fixed 9 merge-introduced regressions found by running the
full Python test suite for the first time on this branch (uninitialized
JPJavaFrame::fast() context, self-vs-type confusion in two PyJPClass_*
helpers, a GIL-released-then-Python-C-API-touched bug in JPMonitor::enter,
a dangling global ref in JPPythonConversion::convert, a dead _JChar
global, JPFunctional's wrong base class dropping pythonConversion for
python.lang.Py* bridge interfaces, a missing match.cacheable=false, and a
JPArray-post-shutdown use-after-free through m_Class->getContext()
surfaced while removing JPContext_global). Restored Support.getJarPath
and its test, silently dropped during an earlier reconciliation pass.

test/jpypetest passes clean (2194 passed, 173 skipped) across multiple
random seeds. The Java testbench (mvn -o test) now boots against this
build when pointed at the venv's Python (-Dpython.executable=...) but
surfaces a separate, pre-existing gap: Python exceptions crossing into
Java via the proxy/hostInvoke path always arrive as the generic
PyExceptionProxy, never the specific python.exceptions.Py*Error subtype
PyExceptionFactory's LOOKUP table implies should be constructed - that
factory is never actually invoked anywhere in the current tree. Left
for a follow-up; not something introduced by this reconciliation.

+++ native/common/jp_convert.cpp
@@ -1,3 +1,4 @@
+// --- file: common/jp_convert.cpp ---
 /*****************************************************************************
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -28,8 +29,8 @@ class Half
 {
 public:
 	static jvalue convert(void* c)
-    {
-        uint16_t i = *(uint16_t*) c;
+	{
+		uint16_t i = *(uint16_t*) c;
 		uint32_t sign = (i&0x8000)>>15;
 		uint32_t exp  = (i&0x7C00)>>10;
 		uint32_t frac = (i&0x03ff);
@@ -45,7 +46,7 @@ public:
 				count = count | (count >> 2);
 				count = count | (count >> 4);
 				count = count | (count >> 8);
-				int zeros = std::bitset<32>(~count).count();
+				int zeros = std::bitset<32>(static_cast<uint32_t>(~count)).count();
 				exp = 127-zeros+7;
 				exp <<= 23;
 				frac <<= zeros-8;
